fix(marketing): remove unused tailwind reference
