+++ packages/marketing/scripts/build.ts
@@ -1,0 +1,6 @@
+import { build } from "../../framework/build";
+
+// Build the marketing site using the framework's build function
+await build();
+
+export {};
