+++ packages/marketing/src/App.tsx
@@ -1,0 +1,12 @@
+import { Button } from "vercel-experiments-ui/components/button";
+
+export function App() {
+  return (
+    <div className="app">
+      <h1 className="text-2xl font-bold">Bun + React</h1>
+      <Button>Click me</Button>
+    </div>
+  );
+}
+
+export default App;
